New rel_ptr implementation
This is necessary to avoid some issues around thread local storage
drop order.

For thread stacks that have an initial value we need to store the
initial value somewhere. That somewhere also needs to be thread local.
If the value is put in a separate thread local from the thread stack
itself rust doesn't provide any guarantee that the thread stack is
destroyed before the initial value. But if they are part of the same
value in the link list based design you end up with a self referential
struct, where the pointer to the top of the stack sometimes is
pointing inside the thread stack itself (Where the initial value is
stored).

We could just use a regular pointer to handle this situation because
thread stacks themselves are never moved so a self referential pointer
would be safe. The problem with that is just the new function
constructing the thread stack and returning it constitutes a move, and
there is currently no way in rust to guarantee a placement new.

So relative pointers seem like the only way. I don't think this should constitute any kind of performance overhead because the assembly should be similar to array indexing but I have not actually scrutinized the generated code.